Put Task related types in a separate file
