+++ algoliasearch/types_responses.go
@@ -16,21 +16,6 @@ type DeleteRes struct {
 	DeletedAt string `json:"deletedAt"`
 }
 
-type DeleteTaskRes struct {
-	DeletedAt string `json:"deletedAt"`
-	TaskID    int64  `json:"taskID"`
-}
-
-type UpdateTaskRes struct {
-	TaskID    int64  `json:"taskID"`
-	UpdatedAt string `json:"updatedAt"`
-}
-
-type TaskStatusRes struct {
-	Status      string `json:"status"`
-	PendingTask bool   `json:"pendingTask"`
-}
-
 type BatchRes struct {
 	ObjectIDs []string `json:"objectIDs"`
 	TaskID    int64    `json:"taskID"`
